# backend/openapi.json
{"openapi":"3.0.0","info":{"title":"News API","description":"A modern backend API for news content","contact":{"name":"Developer Team","url":"","email":""},"license":{"name":"MIT","url":""},"termsOfService":"","version":"0.0.1"},"servers":[{"url":"http://localhost:4000/api","description":"Development server"}],"security":[{"BearerAuth":[]}],"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer"}},"schemas":{"MemoryInfo":{"description":"","type":"object","properties":{"total":{"description":"","type":"string"},"free":{"description":"","type":"string"},"usage":{"description":"","type":"string"}}},"SystemInfo":{"description":"","type":"object","properties":{"platform":{"description":"","type":"string"},"nodeVersion":{"description":"","type":"string"},"memory":{"description":"","$ref":"#/components/schemas/MemoryInfo"},"cpu":{"description":"","type":"string"},"cores":{"description":"","type":"number"}}},"HealthCheck":{"description":"","type":"object","properties":{"status":{"description":"","type":"string"},"timestamp":{"description":"","type":"string"},"uptime":{"description":"","type":"number"},"environment":{"description":"","type":"string"},"system":{"description":"","$ref":"#/components/schemas/SystemInfo"}}},"Disease":{"description":"A Disease","type":"object","properties":{"id":{"description":"Disease identifier","type":"number"},"code":{"description":"ICD-10 code","type":"string"},"name":{"description":"Disease name","type":"string"},"description":{"description":"Detailed description","type":"string"},"prevention":{"description":"Prevention recommendations","type":"string"},"symptoms":{"description":"List of symptoms","type":"array","items":{"type":"string"}},"risks":{"description":"List of risk factors","type":"array","items":{"type":"string"}}}},"RiskFactorDTO":{"description":"A Risk Factor","required":["name"],"type":"object","properties":{"code":{"description":"Risk factor code","type":"string"},"name":{"description":"Risk factor name","type":"string"},"definition":{"description":"Risk factor definition","type":"string"}}},"SymptomDTO":{"description":"A Symptom","required":["term"],"type":"object","properties":{"code":{"description":"SNOMED CT code","type":"string"},"term":{"description":"Symptom term","type":"string"},"category":{"description":"Symptom category","type":"string"}}},"SourceDTO":{"description":"Source Data Transfer Object","type":"object","properties":{"id":{"description":"","type":"number"},"name":{"description":"Source / organization name","type":"string"},"link":{"description":"URL to the data source","type":"string"}}},"ErrorResponse":{"description":"Error Response","required":["message"],"type":"object","properties":{"success":{"description":"Always false for errors","type":"boolean"},"message":{"description":"Human-readable error message","type":"string"},"errors":{"description":"Additional validation errors (optional)","type":"array","items":{"type":"string"}}}}}},"paths":{"/health/details":{"get":{"deprecated":false,"summary":"Detailed health check","security":[],"responses":{"200":{"description":"Detailed health information","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheck"}}}}},"parameters":[],"tags":["Health"]}},"/health":{"get":{"deprecated":false,"summary":"Simple health check","security":[],"responses":{"204":{"description":"No Content"}},"parameters":[],"tags":["Health"]}},"/diseases":{"get":{"deprecated":false,"summary":"Retrieve diseases with pagination and filtering","security":[],"responses":{"200":{"description":"List of diseases","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Disease"}},"examples":{"example1":{"summary":"Success with results","value":[{"id":1,"code":"CAD","name":"Coronary Artery Disease","description":"A condition where the coronary arteries become narrowed or blocked","prevention":"Regular exercise, healthy diet, quit smoking","symptoms":["chest pain","shortness of breath","fatigue"],"risks":["smoking","high cholesterol","diabetes"]}]},"example2":{"summary":"No diseases found","value":{"message":"No diseases found"}}}}}},"400":{"description":"Invalid query parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"examples":{"example3":{"summary":"Invalid skip parameter","value":{"success":false,"message":"Parameter 'skip' must be a non-negative integer"}},"example4":{"summary":"Invalid take parameter","value":{"success":false,"message":"Parameter 'take' must be between 1 and 100"}}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"examples":{"example5":{"summary":"Database connection error","value":{"success":false,"message":"Can't reach database server at localhost:5432"}}}}}}},"parameters":[{"name":"skip","in":"query","description":"Number of records to skip (default: 0, min: 0)","required":false,"deprecated":false,"schema":{"type":"integer"}},{"name":"take","in":"query","description":"Number of records to take (default: 6, min: 1, max: 100)","required":false,"deprecated":false,"schema":{"type":"integer"}},{"name":"symptom","in":"query","description":"Filter by symptom (term or SNOMED code)","required":false,"deprecated":false,"schema":{"type":"string"}},{"name":"riskFactor","in":"query","description":"Filter by risk factor (name or code)","required":false,"deprecated":false,"schema":{"type":"string"}},{"name":"search","in":"query","description":"Free text search across diseases, symptoms, risk factors","required":false,"deprecated":false,"schema":{"type":"string"}},{"name":"locale","in":"query","description":"Language locale (\"en\" or \"ru\", default: \"en\")","required":false,"deprecated":false,"schema":{"type":"string"}}],"tags":["Diseases"]}},"/diseases/risk-factors":{"get":{"deprecated":false,"summary":"Get all risk factors with codes","security":[],"responses":{"200":{"description":"List of risk factors","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RiskFactorDTO"}},"examples":{"example1":{"summary":"Success with results","value":[{"code":"SM","name":"Smoking","definition":"Tobacco use including cigarettes, cigars, and pipes"},{"code":"HG","name":"High cholesterol","definition":"Elevated levels of cholesterol in the blood"}]}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"examples":{"example2":{"summary":"Database error","value":{"success":false,"message":"Failed to fetch risk factors from database"}}}}}}},"parameters":[{"name":"locale","in":"query","description":"Language locale (\"en\" or \"ru\", default: \"en\")","required":false,"deprecated":false,"schema":{"type":"string"}}],"tags":["Diseases"]}},"/diseases/by-letter":{"get":{"deprecated":false,"summary":"Retrieve diseases by initial letter (locale-aware)","security":[],"responses":{"200":{"description":"List of diseases","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Disease"}}}}},"400":{"description":"Invalid query parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"parameters":[{"name":"letter","in":"query","description":"Initial letter to filter by (single character)","required":false,"deprecated":false,"schema":{"type":"string"}},{"name":"skip","in":"query","description":"Number of records to skip (default: 0)","required":false,"deprecated":false,"schema":{"type":"integer"}},{"name":"take","in":"query","description":"Number of records to take (default: 6)","required":false,"deprecated":false,"schema":{"type":"integer"}},{"name":"locale","in":"query","description":"Language locale (\"en\" or \"ru\", default: \"en\")","required":false,"deprecated":false,"schema":{"type":"string"}}],"tags":["Diseases"]}},"/diseases/symptoms":{"get":{"deprecated":false,"summary":"Get all symptoms with codes","security":[],"responses":{"200":{"description":"List of symptoms","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SymptomDTO"}},"examples":{"example1":{"summary":"Success with results","value":[{"code":"CP","term":"Chest pain","category":"Cardiovascular"},{"code":"SOB","term":"Shortness of breath","category":"Respiratory"}]}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"examples":{"example2":{"summary":"Database error","value":{"success":false,"message":"Failed to fetch symptoms from database"}}}}}}},"parameters":[{"name":"locale","in":"query","description":"Language locale (\"en\" or \"ru\", default: \"en\")","required":false,"deprecated":false,"schema":{"type":"string"}}],"tags":["Diseases"]}},"/":{"get":{"deprecated":false,"summary":"Get API information","security":[],"responses":{"200":{"description":"API info","content":{"application/json":{"schema":{"type":"object"}}}}},"parameters":[],"tags":["API"]}},"/debug-sentry":{"get":{"deprecated":false,"summary":"Debug endpoint for Sentry testing","security":[],"responses":{"500":{"description":"Internal server error (intentional)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}},"parameters":[],"tags":["API"]}},"/sources":{"get":{"deprecated":false,"summary":"Get sources with pagination and optional search","security":[],"responses":{"200":{"description":"List of sources","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SourceDTO"}},"examples":{"example1":{"summary":"Success with results","value":[{"id":1,"name":"World Health Organization","link":"https://www.who.int"},{"id":2,"name":"American Heart Association","link":"https://www.heart.org"}]},"example2":{"summary":"No sources found","value":{"message":"No sources found"}}}}}},"400":{"description":"Invalid query parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"examples":{"example3":{"summary":"Invalid skip parameter","value":{"success":false,"message":"Parameter 'skip' must be a non-negative integer"}},"example4":{"summary":"Invalid take parameter","value":{"success":false,"message":"Parameter 'take' must be between 1 and 100"}}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"},"examples":{"example5":{"summary":"Database connection error","value":{"success":false,"message":"Can't reach database server at localhost:5432","stack":"PrismaClientInitializationError: ..."}}}}}}},"parameters":[{"name":"skip","in":"query","description":"Number of records to skip (default: 0, min: 0)","required":false,"deprecated":false,"schema":{"type":"integer"}},{"name":"take","in":"query","description":"Number of records to take (default: 6, min: 1, max: 100)","required":false,"deprecated":false,"schema":{"type":"integer"}},{"name":"search","in":"query","description":"Search by source / organization name","required":false,"deprecated":false,"schema":{"type":"string"}}],"tags":["Sources"]}}},"tags":[{"name":"API","description":""},{"name":"Diseases","description":""},{"name":"Health","description":""},{"name":"Sources","description":""}]}